+++ src/convertFormats.cpp
@@ -89,15 +89,15 @@ size_t findEllCols(const COO<ValueType> &coo, const size_t block_size) {
 }
 
 template <typename ValueType>
-std::unique_ptr<size_t[]> findColBlockIdx(const COO<ValueType> &coo,
-                                          const size_t block_size,
-                                          const size_t ell_cols,
-                                          const size_t padded_rows) {
+std::unique_ptr<int[]> findColBlockIdx(const COO<ValueType> &coo,
+                                       const size_t block_size,
+                                       const size_t ell_cols,
+                                       const size_t padded_rows) {
   const size_t nblock_rows = padded_rows / block_size;
   const size_t nblock_cols = ell_cols / block_size;
   const size_t nblocks = nblock_rows * nblock_cols;
 
-  auto col_block_idx = std::make_unique<size_t[]>(nblocks);
+  auto col_block_idx = std::make_unique<int[]>(nblocks);
   std::fill(col_block_idx.get(), col_block_idx.get() + nblocks, -1);
 
   size_t current_block_row = -1;
@@ -146,7 +146,7 @@ std::unique_ptr<size_t[]> findColBlockIdx(const COO<ValueType> &coo,
 template <typename ValueType>
 std::unique_ptr<ValueType[]> findVals(
     const COO<ValueType> &coo, const size_t block_size, const size_t ell_cols,
-    const size_t padded_rows, const std::unique_ptr<size_t[]> &col_block_idx) {
+    const size_t padded_rows, const std::unique_ptr<int[]> &col_block_idx) {
   const size_t nblock_cols = ell_cols / block_size;
   auto vals = std::make_unique<ValueType[]>(padded_rows * ell_cols);
   std::fill(vals.get(), vals.get() + padded_rows * ell_cols, 0);
